contract-engine/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-docx
docxtpl
jinja2
lxml